Network: . Added a singleton EntityMapper class which automatically reads & writes object diffs . All entities are translated properly
Editor:
. No longer blurs the edges of pixel art

Client:
. Reverted smart movement detection as it would glitch using network
. Store last played multiplayer IP
. Renamed UI.Counter to UI.GlowTimer

Common:
. Clients can now send a disconnect message to the server
. Entity.Scale is now a float
. Stub objects now implement their ObjectType directly

Engine:
. Fix a bug which made all units able to fly
. Updated existing and added new coherent noise primitives

+++ Editor/Views/Maps/PropPanel.cs
@@ -20,7 +20,7 @@ namespace Shanism.Editor.Views.Maps
             TypeName = DoodadType,
             Model = string.Empty,
             Owner = string.Empty,
-            Size = 2.5,
+            Size = 2.5f,
             Tint = Common.Color.White,
         };
 
@@ -80,7 +80,7 @@ namespace Shanism.Editor.Views.Maps
 
         void numericUpDown1_ValueChanged(object sender, EventArgs e)
         {
-            constr.Size = (double)btnSize.Value;
+            constr.Size = (float)btnSize.Value;
 
             BrushChanged?.Invoke(constr);
         }
